Fix #16: Splitting into HTML files
- Add nice DIFFing of unit test veryfing info file
- Verify paragraph title of a page
- Extract PDF and EPUB processing into separate classes
- Add configuration object to main class
- Expand info file with configuration details
- Make configuration a singleton available from everywhere
- Support skipping pages listed in skip section of configuration as list
  or arrays. Index is the same as page number of a PDF document.
- Expose capturing PDF region into configuration
- Remove unnecessary, empty entries
- Move test files to 'accident' directory, enable test for chapters
