+++ src/test/resources/accident/convert.json
@@ -1,0 +1,13 @@
+{
+  "pdfFilePath" : "src/test/resources/accident/original.pdf",
+  "epubFilePath" : "target/accident_ebook.epub",
+  "skip": {
+    "arrays": [[1, 8], [179, 227]]
+  },
+  "region": {
+    "x": 0,
+    "y": 80,
+    "weight": 550,
+    "height": 520
+  }
+}
