--- src/test/resources/ebook.epub.info
@@ -1,84 +1,0 @@
-Doc info: Title  : Accident: The Death of General Sikorski
-Doc info: Author : David Irving
-Doc info: Subject: Accident
-Doc info: Pages  : 227
-Structure: Chapter 1 :  “Soldiers Must Die” 9
-Structure: Chapter 2 :  Six weeks too Soon 39
-Structure: Chapter 3 :  Farce and Tragedy 51
-Structure: Chapter 4 :  Search and Inquire 69
-Structure: Chapter 5 :  Mr Churchill Kneels in Prayer 101
-Structure: Chapter 6 :  Mailbags and Manifests 105
-Structure: Chapter 7 :  The Unmentioned Issue 129
-Structure: Chapter 8 :  Post Mortem 153
-Structure: Chapter 9 :  Open Verdict 153
-Structure: Chapter 1 :   “Soldiers Must Die” 163
-Structure: Chapter 2 :   Six Weeks too Soon 170
-Structure: Chapter 3 :  Farce and Tragedy 172
-Structure: Chapter 4 :  Search  and Inquire 176
-Structure: Chapter 5 :  Mr Churchill Kneels in Prayer 187
-Structure: Chapter 6 :  Mailbags and Manifests 188
-Structure: Chapter 7 :  The Unmentioned Issue 194
-Structure: Chapter 8 :  Postmortem 200
-Structure: Chapter 9 :  Open Verdict 203
-Structure: Chapter 1 :  “Soldiers Must Die”
-Structure: Chapter 2 :  Six weeks too Soon
-Structure: Part TWO : 
-Structure: Chapter 3 :  Farce and Tragedy
-Structure: Chapter 4 :  Search and Inquire
-Structure: Chapter 5 :  Mr Churchill Kneels in
-Structure: Chapter 6 :  Mailbags and Manifests
-Structure: Chapter 7 :  The Unmentioned Issue
-Structure: Chapter 8 :  Post Mortem
-Structure: Chapter 9 :  Open Verdict
-Structure: Chapter 1 :   “Soldiers Must Die”
-Structure: Chapter 2 :   Six Weeks too Soon
-Structure: Chapter 3 :  Farce and Tragedy
-Structure: Chapter 4 :  Search  and Inquire
-Structure: Chapter 58 :  “Broken cigarette cartons were tinging the water green.” It must
-Structure: Chapter 5 :  Mr Churchill Kneels in Prayer
-Structure: Chapter 6 :  Mailbags and Manifests
-Structure: Chapter 7 :  The Unmentioned Issue
-Structure: Chapter 8 :  Postmortem
-Structure: Chapter 9 :  Open Verdict
-Structure: Chapter 1943 :  “Both main engine ignition switches were found off.”
-Contains 40 pages
-META-INF/container.xml => 237 lines, hash=-1852420702
-content.opf => 4948 lines, hash=713717394
-content1.html => 1439 lines, hash=120190586
-content10.html => 544 lines, hash=-852925143
-content11.html => 534 lines, hash=1991371292
-content12.html => 529 lines, hash=596914197
-content13.html => 527 lines, hash=-1947140845
-content14.html => 529 lines, hash=1239855900
-content15.html => 539 lines, hash=-1951867788
-content16.html => 532 lines, hash=1462106276
-content17.html => 531 lines, hash=414745021
-content18.html => 520 lines, hash=-1764355181
-content19.html => 3114 lines, hash=-2083884751
-content2.html => 531 lines, hash=1374834924
-content20.html => 63890 lines, hash=696209310
-content21.html => 20851 lines, hash=-1318132452
-content22.html => 548 lines, hash=726729567
-content23.html => 9405 lines, hash=-1282419154
-content24.html => 73153 lines, hash=1043227969
-content25.html => 7182 lines, hash=2036935602
-content26.html => 14380 lines, hash=610983305
-content27.html => 14498 lines, hash=2083235096
-content28.html => 33671 lines, hash=1787408381
-content29.html => 3464 lines, hash=-979695871
-content3.html => 556 lines, hash=160489806
-content30.html => 16370 lines, hash=1482388742
-content31.html => 6434 lines, hash=-1937601505
-content32.html => 16234 lines, hash=-423375972
-content33.html => 10205 lines, hash=471615553
-content34.html => 2355 lines, hash=214373621
-content35.html => 7844 lines, hash=1396808743
-content36.html => 14752 lines, hash=-376901289
-content4.html => 526 lines, hash=2117962545
-content5.html => 527 lines, hash=1672418469
-content6.html => 539 lines, hash=1287346475
-content7.html => 576 lines, hash=-720683913
-content8.html => 531 lines, hash=-700151765
-content9.html => 521 lines, hash=-445169479
-mimetype => 20 lines, hash=-1000808489
-toc.ncx => 7222 lines, hash=-796221706
